requirements.txt
fastapi[All]
pytest
hypothesis
pytest-asyncio
pytest-mock
python-jose[cryptography]
aio_pika
pydantic
aioredis
coloredlogs
fakeredis
black
flake8
mypy